create a simple file storage
This allows to upload the picture on a the file system if an Azure storage isn't configured.

This is a naïve implementation which follows `azure-storage-node` API in order to minimize the impact on the CBoard API codebase.
More work should be done to secure the endpoint and only allow images to be uploaded.